MapperEngine, Mapper/Phase, Confre, and AutoArr mapper now all work
MapperEngine deserializes UPPAAL models into "UppaalPojo" objects and traverses this tree. All nodes are sent to mapper-phases for "handling".

Each "Mapper" has multiple "Phase"s that can register various handlers on "UppaalPojo" types. Thus, whenever a UPPAAL object (e.g. declaration in template) is hit, the current phase may (not) have a handler to mutate this object.

MapperEngines can contain multiple mappers.

The "AutoArr" mapper replaces all "{ x }" array initializers with complete initializers based on the size of the array (only works with integers (not variables) for now).

The "Confre" is a "Regex", but the pattern is a context-free grammar and the output of a "match" is a tree.