SeComp: Added support for using partial inst. on system line AND as base
In case a partial instantiation is used as the base for another partial instantiation, a non-free parameter is added to the base-partial-inst.

Thus, if said base is also used on the system line, SeComp now generates a new partial instantiation without the non-free parameter to use on the system line instead.

The name of the resulting auto-generated process(es) is mapped back to the original base's name to hide this transformation in the simulator.